revert to int

+++ project_3/graph.py
@@ -10,7 +10,7 @@ height = []
 with open('myfile.txt', 'rt') as infile:
     lines = infile.read().split(',')
 for i in range(0, len(lines) - 1):
-    height.append(float(lines[i]))
+    height.append(int(lines[i]))
 print(height)
 bars = ('FIFO', 'SJN', 'PRIORITY', 'RR - 10', 'RR - 100', 'RR - 1000', 'RR(SRTN)', 'RR(Priority)')
 y_pos = range(len(bars))
